// src/types.ts
export type CategoryId = 
  | 'visiting-cards'
  | 'business-cards'
  | 'letterheads'
  | 'envelopes'
  | 'invitation-cards'
  | 'brochures'
  | 'paper-board-files'
  | 'project-files'
  | 'flyers'
  | 'stickers'
  | 'packaging'
  | 'banners'
  | 'id-cards'
  | 'stationery'
  | 'custom-merch'
  | 'apparel'
  | 'menus'
  | 'all'
  | string;